+++ common/src/main/java/dev/detpikachu/unpluggedafk/common/formatting/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package dev.detpikachu.unpluggedafk.common.formatting;
+
+import org.jspecify.annotations.NullMarked;